Credits no longer have to be integer (credits arranged to be double)

+++ lib/screens/combination_calculation/combinations_shower.dart
@@ -30,7 +30,7 @@ class _CombinationsShowerState extends State<CombinationsShower> {
   List<Lecture> sortedLectures = Term.instance.sortedLectures;
   double? selectedGpa;
   List<int> indexes = [];
-  int totCred = 0;
+  double totCred = 0;
   List<List<double>> effects = [];
   bool showDetails = true;
 
@@ -127,7 +127,7 @@ class _CombinationsShowerState extends State<CombinationsShower> {
                                   for (int i = 0; i < indexes.length; i++) {
                                     final comb = combinations[indexes[i]];
 
-                                    int cred = actcred;
+                                    double cred = actcred;
                                     double smp = actgpa * actcred;
                                     effects.add([]);
                                     for (
@@ -603,7 +603,7 @@ class _CombinationsShowerState extends State<CombinationsShower> {
       "${t(context).term}\t${t(context).gpa_column}\t${t(context).credit}\n",
     );
     double oldgpa = Term.instance.oldgpa;
-    int oldcred = Term.instance.oldcred;
+    double oldcred = Term.instance.oldcred;
     double diff = ((selectedGpa! - oldgpa) * 100).round() / 100;
     String diffgpa = diff < 0 ? '$diff' : '+$diff';
     String diffcred = totCred - oldcred < 0
